updating licenses
We are shifting from a cc-sa to cc-nc-sa for the non-code content. For the code content, we will be using a GPL 3 license but noting that those unable to use a GPL license can reach out to explore dual licensing under a compatible license. Also changed the folder structure to clarify which license applies to which content and documented this in the readme.